feat(nav): grouped IA — Work/Settings sections, Skills folds into Library
- layout.tsx: ~6 primary entries (Home · Work[Tasks/Projects/Decisions] ·
  Library · Codebases · Query · Settings[Team tools/Admin])
- team-nav.tsx: NavSection groups (non-clickable headers + indented leaves)
- move /skills → /library/skills (canonical view); redirect stub at old path
- Library 'skill' kind-chip links to /library/skills (single skills surface)
